Extract Router
